Changing docker build scripts into a templating engine
The ROC infrastructure has a few permutations that are useful
in different circumstances.  Namely, the branch to build source from
and the build configuration.  Instead of creating seperate dockerfiles
for each variant, which could explode, the setup script has been
changed into a templating engine, wherin it transforms a docker template
file into a real dockerfile by substituting keywards.  It is similar
to a pre-processor for the dockerfiles.